Allow creating a dummy recovery patch file.
Since we don't make use of the recovery patch file, it's a waste of space.
However, we can't completely get rid of the file without opening a can of worms
in releasetools, among other things. This patch allows replacing it with a
blank file by setting TARGET_DUMMY_RECOVERY_PATCH to true.